+++ esp-hal/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `AnyPin` now implements `From<GpioPin<N>>`. (#2326)
 
+- `Pins::steal()` to unsafely obtain GPIO. (#2335)
+
 ### Changed
 
 ### Fixed